+++ src/compas_cadwork/elements/dimensional_element.py
@@ -1,7 +1,9 @@
 import geometry_controller as gc
+import visualization_controller as vc
 from compas.geometry import Line
 
 from compas_cadwork.conversions.primitives import point_to_compas
+from compas_cadwork.material import Material
 
 from .element import T
 from .oriented_element import OrientedElement
@@ -43,3 +45,15 @@ class DimensionalElement(OrientedElement[T]):
         p1 = point_to_compas(gc.get_p1(self.id))
         p2 = point_to_compas(gc.get_p2(self.id))
         return Line(p1, p2)
+
+    @property
+    def material(self) -> Material:
+        """Element material."""
+        material_id = vc.get_material(self.id)
+        if material_id == 0:
+            raise RuntimeError(f"Cadwork element #{self.id} no longer exists")
+        return Material(material_id)
+
+    @material.setter
+    def material(self, value: Material) -> None:
+        vc.set_material([self.id], value.id)