+++ src/compas_cadwork/project.py
@@ -11,6 +11,7 @@ from uuid import UUID
 
 import bim_controller as bc
 import element_controller as ec
+import material_controller as mc
 import utility_controller as uc
 
 from compas_cadwork.elements.factory import AnyElement
@@ -19,9 +20,12 @@ from compas_cadwork.ifc_uuid import IfcUUID
 from compas_cadwork.utils.storage import IterableKeyValueStorage
 from compas_cadwork.utils.storage import KeyValueStorage
 
+from .material import Material
+
 
 if TYPE_CHECKING:
     from cadwork import ElementId
+    from cadwork import MaterialId
 
 
 def _is_empty_value(raw_value: str) -> bool:
@@ -308,6 +312,75 @@ class Project:
         """
         for cadwork_id in ec.get_active_identifiable_element_ids():
             yield get_element_instance(cadwork_id)
+
+    @overload
+    def material(self, *, cadwork_id: MaterialId) -> Material:
+        """Get material from Cadwork ID.
+
+        Parameters
+        ----------
+        cadwork_id : MaterialId
+            Cadwork material ID.
+
+        Returns
+        -------
+        Material
+            Cadwork material.
+
+        Raises
+        ------
+        ValueError
+            If the material does not exist.
+        """
+
+    @overload
+    def material(self, *, name: str) -> Material:
+        """Get material from name.
+
+        Parameters
+        ----------
+        name: str
+            Material name.
+
+        Returns
+        -------
+        Material
+            Cadwork material.
+
+        Raises
+        ------
+        ValueError
+            If the material does not exist.
+        """
+
+    def material(
+        self,
+        *,
+        cadwork_id: MaterialId | None = None,
+        name: str | None = None,
+    ):
+        # Name to Cadwork ID
+        if name is not None:
+            cadwork_id = mc.get_material_id(name)
+            if cadwork_id == 0:
+                raise ValueError(f"Could not find a Cadwork material with name {name!r}")
+
+        # Cadwork ID to material
+        assert cadwork_id is not None
+        if mc.get_name(cadwork_id) == "":
+            raise ValueError(f"Could not find a Cadwork material with ID #{cadwork_id}")
+        return Material(cadwork_id)
+
+    def materials(self) -> Generator[Material, None, None]:
+        """Get all materials in the project.
+
+        Returns
+        -------
+        Generator[Material, None, None]
+            Generator of materials.
+        """
+        for material_id in mc.get_all_materials():
+            yield Material(material_id)
 
     def __repr__(self) -> str:
         return f"Project(guid={self.guid!r}, name={self.name!r})"