+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "do not merge" workflow check to CI/CD pipeline.
 - Added `ElementType` enumeration to be used with `Element.type` property.
 - Added initial `Beam`, `Panel`, and `Wall` classes.
+- Added support for materials through the `Material` class.
 
 ### Changed